_evaluator/requirements.txt
bs4
lxml
Pillow